+++ GameEngine/GameEngine/GoToPointAI.cpp
@@ -22,8 +22,8 @@ bool GoToPointAI::execute(LivingObject * obj, float dt)
 		return true;
 	}
 	else {
-		int dx = dest[0] - obj->pos[0];
-		int dy = dest[1] - obj->pos[1];
+		float dx = dest[0] - obj->pos[0];
+		float dy = dest[1] - obj->pos[1];
 		float angle = atan2(dy,dx);
 		obj->pos[0] += cos(angle) * obj->moveSpeed * dt;
 		obj->pos[1] += sin(angle) * obj->moveSpeed * dt;
